Extract speech generation core
